% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoid_enclave.R
\name{avoid_enclave}
\alias{avoid_enclave}
\title{Merge Municipalities to Group Together To Avoid Creating Enclaves}
\usage{
avoid_enclave(pref, merge_codes)
}
\arguments{
\item{pref}{sf object of cleaned, collated, census data for a prefecture}

\item{merge_codes}{an array of codes for municipalities to group together so as to avoid creating enclaves}
}
\value{
an sf object with the chosen municipalities grouped together
}
\description{
Merge Municipalities to Group Together To Avoid Creating Enclaves
}
\concept{getdata}
