% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_jcdf.R
\name{clean_jcdf}
\alias{clean_jcdf}
\title{Download JCDF (Japan Census Data File)}
\usage{
clean_jcdf(path)
}
\arguments{
\item{path}{string containing path to the shapefile census data}
}
\value{
a sf object with cleaned, collated version of the census data
}
\description{
Download JCDF (Japan Census Data File)
}
\concept{getdata}
