% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pop_demographics.R
\name{download_pop_demographics}
\alias{download_pop_demographics}
\title{Download Prefectural Population Demographics Data}
\usage{
download_pop_demographics(pref_code)
}
\arguments{
\item{pref_code}{administrative code for prefecture of choice (eg. Hokkaido: 01, Okinawa: 47)}
}
\value{
demographic data downloaded in data/ directory and resulting csv
}
\description{
Download Prefectural Population Demographics Data
}
\concept{downloaddata}
