% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municodes.R
\name{get_municodes}
\alias{get_municodes}
\title{Get Municipality Codes}
\usage{
get_municodes(pref_raw)
}
\arguments{
\item{pref_raw}{sf object of census data for a prefecture}
}
\value{
an array with the municipality name and corresponding codes
}
\description{
Get Municipality Codes
}
\concept{getdata}
