% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_small.R
\name{merge_small}
\alias{merge_small}
\title{Merge Components of Municipalities of Choice}
\usage{
merge_small(pref, split_codes)
}
\arguments{
\item{pref}{sf object of cleaned, collated, census data for a prefecture}

\item{split_codes}{an array of codes for municipalities to split}
}
\value{
an sf object with the chosen municipalities split
}
\description{
Merge Components of Municipalities of Choice
}
\concept{getdata}
