% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_kakusa.R
\name{original_kakusa}
\alias{original_kakusa}
\title{Calculate Original Ippyo-no-kakusa (maximum weight disparity in voter power)}
\usage{
original_kakusa(pref)
}
\arguments{
\item{pref}{an sf object with cleaned, collated version of the census data}
}
\value{
a vector with the ippyo-no-kakusa for the oiginal plan
}
\description{
Calculate Original Ippyo-no-kakusa (maximum weight disparity in voter power)
}
\concept{calculate}
