% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_weight_disparity_table.R
\name{original_weight_disparity_table}
\alias{original_weight_disparity_table}
\title{Calculate Original Ippyo-no-kakusa (maximum weight disparity in voter power)}
\usage{
original_weight_disparity_table(pref)
}
\arguments{
\item{pref}{an sf object with cleaned, collated version of the census data}
}
\value{
a table with the ippyo no kakusa parameters for the original plan
}
\description{
Calculate Original Ippyo-no-kakusa (maximum weight disparity in voter power)
}
\concept{calculate}
