% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_maxmin_LH_table.R
\name{simulation_maxmin_LH_table}
\alias{simulation_maxmin_LH_table}
\title{Calculate Ippyo-no-kakusa of simulation output(maximum weight disparity in voter power)}
\usage{
simulation_maxmin_LH_table(redist_simulation_output)
}
\arguments{
\item{redist_simulation_output}{a data table of the \code{redist} simulation function output (\code{redist_smc} etc.)}
}
\value{
a table (n_sims x 2) with the ippyo no kakusa parameters for the simulated plan

a table (n_sims x 4) with the ippyo no kakusa parameters for the simulated plan
}
\description{
Calculate Ippyo-no-kakusa of simulation output(maximum weight disparity in voter power)
}
\concept{calculate}
