% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lake.R
\name{remove_lake}
\alias{remove_lake}
\title{Remove a Lake}
\usage{
remove_lake(pref_raw, sf_lake, name_lake)
}
\arguments{
\item{pref_raw}{sf object of census data for a prefecture}

\item{sf_lake}{sf object of GIS data of Japanese lakes}

\item{name_lake}{character vector containing name of lakes}
}
\value{
an array with the municipality name and corresponding codes
}
\description{
Remove a Lake
}
\concept{getdata}
