% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_quo_match.R
\name{status_quo_match}
\alias{status_quo_match}
\title{Match current districts}
\usage{
status_quo_match(pref)
}
\arguments{
\item{pref}{sf object of cleaned, collated, census data for a prefecture}
}
\value{
sf object with current legislative district code added
}
\description{
Match current districts
}
\concept{getdata}
